rename files and directory
